Split rpc blueprint
