+++ starknet_devnet/blueprints/rpc/structures/payloads.py
@@ -1,7 +1,5 @@
 """
-RPC utilities
-API Specification v0.1.0
-https://github.com/starkware-libs/starknet-specs/releases/tag/v0.1.0
+RPC payload structures
 """
 
 from __future__ import annotations
@@ -11,12 +9,9 @@ from typing import Callable, Union, List, Optional
 from starkware.starknet.definitions.general_config import StarknetGeneralConfig
 from starkware.starknet.services.api.contract_class import ContractClass
 from starkware.starknet.services.api.feeder_gateway.response_objects import (
-    BlockStatus,
     StarknetBlock,
     InvokeSpecificInfo,
     DeploySpecificInfo,
-    TransactionReceipt,
-    TransactionStatus,
     TransactionSpecificInfo,
     TransactionType,
     BlockStateUpdate,
@@ -24,74 +19,12 @@ from starkware.starknet.services.api.feeder_gateway.response_objects import (
 )
 from starkware.starknet.services.api.gateway.transaction import InvokeFunction
 from starkware.starknet.services.api.gateway.transaction_utils import compress_program
-from typing_extensions import TypedDict, Literal
+from typing_extensions import TypedDict
 
+from starknet_devnet.blueprints.rpc.rpc_utils import rpc_root, rpc_felt
+from starknet_devnet.blueprints.rpc.structures.types import RpcBlockStatus, BlockHash, BlockNumber, Felt, \
+    rpc_block_status, TxnHash, Address, NumAsHex, TxnType, rpc_txn_type
 from starknet_devnet.state import state
-from ..util import StarknetDevnetException
-
-Felt = str
-
-BlockHash = Felt
-BlockNumber = int
-BlockTag = Literal["latest", "pending"]
-
-
-class BlockHashDict(TypedDict):
-    """
-    TypedDict class for BlockId with block hash
-    """
-    block_hash: BlockHash
-
-
-class BlockNumberDict(TypedDict):
-    """
-    TypedDict class for BlockId with block number
-    """
-    block_number: BlockNumber
-
-
-BlockId = Union[BlockHashDict, BlockNumberDict, BlockTag]
-
-TxnStatus = Literal["PENDING", "ACCEPTED_ON_L2", "ACCEPTED_ON_L1", "REJECTED"]
-
-RpcBlockStatus = Literal["PENDING", "ACCEPTED_ON_L2", "ACCEPTED_ON_L1", "REJECTED"]
-
-
-def rpc_block_status(block_status: BlockStatus) -> RpcBlockStatus:
-    """
-    Convert gateway BlockStatus to RpcBlockStatus
-    """
-    block_status_map = {
-        "PENDING": "PENDING",
-        "ABORTED": "REJECTED",
-        "REVERTED": "REJECTED",
-        "ACCEPTED_ON_L2": "ACCEPTED_ON_L2",
-        "ACCEPTED_ON_L1": "ACCEPTED_ON_L1"
-    }
-    return block_status_map[block_status]
-
-
-TxnHash = Felt
-Address = Felt
-NumAsHex = str
-
-# Pending transactions will not be supported since it
-# doesn't make much sense with the current implementation of devnet
-TxnType = Literal["DECLARE", "DEPLOY", "INVOKE"]
-
-
-def rpc_txn_type(transaction_type: str) -> TxnType:
-    """
-    Convert gateway transaction type to RPC TxnType
-    """
-    txn_type_map = {
-        "DEPLOY": "DEPLOY",
-        "DECLARE": "DECLARE",
-        "INVOKE_FUNCTION": "INVOKE",
-    }
-    if transaction_type not in txn_type_map:
-        raise RpcError(code=-1, message=f"Current implementation does not support {transaction_type} transaction type")
-    return txn_type_map[transaction_type]
 
 
 class RpcBlock(TypedDict):
@@ -125,8 +58,8 @@ async def rpc_block(block: StarknetBlock, tx_type: Optional[str] = "TXN_HASH") -
 
     def config() -> StarknetGeneralConfig:
         devnet_state = state.starknet_wrapper.get_state()
-        config = devnet_state.general_config
-        return config
+        _config = devnet_state.general_config
+        return _config
 
     mapping: dict[str, Callable] = {
         "TXN_HASH": transaction_hashes,
@@ -455,306 +388,3 @@ def rpc_state_update(state_update: BlockStateUpdate) -> RpcStateUpdate:
         }
     }
     return rpc_state
-
-
-class RpcInvokeTransactionResult(TypedDict):
-    """
-    TypedDict for rpc invoke transaction result
-    """
-    transaction_hash: TxnHash
-
-
-class RpcDeclareTransactionResult(TypedDict):
-    """
-    TypedDict for rpc declare transaction result
-    """
-    transaction_hash: TxnHash
-    class_hash: Felt
-
-
-class RpcDeployTransactionResult(TypedDict):
-    """
-    TypedDict for rpc deploy transaction result
-    """
-    transaction_hash: TxnHash
-    contract_address: Felt
-
-
-class MessageToL1(TypedDict):
-    """
-    TypedDict for rpc message from l2 to l1
-    """
-    to_address: Felt
-    payload: List[Felt]
-
-
-class MessageToL2(TypedDict):
-    """
-    TypedDict for rpc message from l1 to l2
-    """
-    from_address: str
-    payload: List[Felt]
-
-
-class Event(TypedDict):
-    """
-    TypedDict for rpc event
-    """
-    from_address: Address
-    keys: List[Felt]
-    data: List[Felt]
-
-
-class RpcBaseTransactionReceipt(TypedDict):
-    """
-    TypedDict for rpc transaction receipt
-    """
-    # Common
-    transaction_hash: TxnHash
-    actual_fee: Felt
-    status: TxnStatus
-    status_data: Optional[str]
-    block_hash: BlockHash
-    block_number: BlockNumber
-
-
-class RpcInvokeReceipt(TypedDict):
-    """
-    TypedDict for rpc invoke transaction receipt
-    """
-    messages_sent: List[MessageToL1]
-    l1_origin_message: Optional[MessageToL2]
-    events: List[Event]
-    # Common
-    transaction_hash: TxnHash
-    actual_fee: Felt
-    status: TxnStatus
-    status_data: Optional[str]
-    block_hash: BlockHash
-    block_number: BlockNumber
-
-
-class RpcDeclareReceipt(TypedDict):
-    """
-    TypedDict for rpc declare transaction receipt
-    """
-    # Common
-    transaction_hash: TxnHash
-    actual_fee: Felt
-    status: TxnStatus
-    status_data: Optional[str]
-    block_hash: BlockHash
-    block_number: BlockNumber
-
-
-class RpcDeployReceipt(TypedDict):
-    """
-    TypedDict for rpc declare transaction receipt
-    """
-    # Common
-    transaction_hash: TxnHash
-    actual_fee: Felt
-    status: TxnStatus
-    status_data: Optional[str]
-    block_hash: BlockHash
-    block_number: BlockNumber
-
-
-def rpc_invoke_receipt(txr: TransactionReceipt) -> RpcInvokeReceipt:
-    """
-    Convert rpc invoke transaction receipt to rpc format
-    """
-    def l2_to_l1_messages() -> List[MessageToL1]:
-        messages = []
-        for message in txr.l2_to_l1_messages:
-            msg: MessageToL1 = {
-                "to_address": rpc_felt(message.to_address),
-                "payload": [rpc_felt(p) for p in message.payload]
-            }
-            messages.append(msg)
-        return messages
-
-    def l1_to_l2_message() -> Optional[MessageToL2]:
-        if txr.l1_to_l2_consumed_message is None:
-            return None
-
-        msg: MessageToL2 = {
-            "from_address": txr.l1_to_l2_consumed_message.from_address,
-            "payload": [rpc_felt(p) for p in txr.l1_to_l2_consumed_message.payload]
-        }
-        return msg
-
-    def events() -> List[Event]:
-        _events = []
-        for event in txr.events:
-            event: Event = {
-                "from_address": rpc_felt(event.from_address),
-                "keys": [rpc_felt(e) for e in event.keys],
-                "data": [rpc_felt(d) for d in event.data],
-            }
-            _events.append(event)
-        return _events
-
-    base_receipt = rpc_base_transaction_receipt(txr)
-    receipt: RpcInvokeReceipt = {
-        "messages_sent": l2_to_l1_messages(),
-        "l1_origin_message": l1_to_l2_message(),
-        "events": events(),
-        **base_receipt,
-    }
-    return receipt
-
-
-def rpc_declare_receipt(txr) -> RpcDeclareReceipt:
-    """
-    Convert rpc declare transaction receipt to rpc format
-    """
-    return rpc_base_transaction_receipt(txr)
-
-
-def rpc_deploy_receipt(txr) -> RpcDeployReceipt:
-    """
-    Convert rpc deploy transaction receipt to rpc format
-    """
-    return rpc_base_transaction_receipt(txr)
-
-
-def rpc_base_transaction_receipt(txr: TransactionReceipt) -> RpcBaseTransactionReceipt:
-    """
-    Convert gateway transaction receipt to rpc transaction receipt
-    """
-    def status() -> str:
-        if txr.status is None:
-            return "UNKNOWN"
-
-        mapping = {
-            TransactionStatus.NOT_RECEIVED: "UNKNOWN",
-            TransactionStatus.ACCEPTED_ON_L2: "ACCEPTED_ON_L2",
-            TransactionStatus.ACCEPTED_ON_L1: "ACCEPTED_ON_L1",
-            TransactionStatus.RECEIVED: "RECEIVED",
-            TransactionStatus.PENDING: "PENDING",
-            TransactionStatus.REJECTED: "REJECTED",
-        }
-        return mapping[txr.status]
-
-    def status_data() -> Union[str, None]:
-        if txr.transaction_failure_reason is not None:
-            if txr.transaction_failure_reason.error_message is not None:
-                return txr.transaction_failure_reason.error_message
-        return None
-
-    receipt: RpcBaseTransactionReceipt = {
-        "transaction_hash": rpc_felt(txr.transaction_hash),
-        "actual_fee": rpc_felt(txr.actual_fee or 0),
-        "status": status(),
-        "status_data": status_data(),
-        "block_hash": rpc_felt(txr.block_hash) if txr.block_hash is not None else None,
-        "block_number": txr.block_number,
-    }
-    return receipt
-
-
-def rpc_transaction_receipt(txr: TransactionReceipt) -> dict:
-    """
-    Convert gateway transaction receipt to rpc format
-    """
-    tx_mapping = {
-        TransactionType.DEPLOY: rpc_deploy_receipt,
-        TransactionType.INVOKE_FUNCTION: rpc_invoke_receipt,
-        TransactionType.DECLARE: rpc_declare_receipt,
-    }
-    transaction = state.starknet_wrapper.transactions.get_transaction(hex(txr.transaction_hash)).transaction
-    tx_type = transaction.tx_type
-    return tx_mapping[tx_type](txr)
-
-
-def block_tag_to_block_number(block_id: BlockId) -> BlockId:
-    """
-    Changes block_id from tag to dict with "block_number" field
-    """
-    if isinstance(block_id, str):
-        if block_id == "latest":
-            return {"block_number": state.starknet_wrapper.blocks.get_number_of_blocks() - 1}
-
-        if block_id == "pending":
-            raise RpcError(code=-1, message="Calls with block_id == 'pending' are not supported currently.")
-
-        raise RpcError(code=24, message="Invalid block id")
-
-    return block_id
-
-
-def get_block_by_block_id(block_id: BlockId) -> dict:
-    """
-    Get block using different method depending on block_id type
-    """
-    if block_id in ["latest", "pending"]:
-        block_id = block_tag_to_block_number(block_id)
-
-    try:
-        if "block_hash" in block_id:
-            return state.starknet_wrapper.blocks.get_by_hash(block_hash=block_id["block_hash"])
-        return state.starknet_wrapper.blocks.get_by_number(block_number=block_id["block_number"])
-    except StarknetDevnetException as ex:
-        raise RpcError(code=24, message="Invalid block id") from ex
-
-
-def assert_block_id_is_latest(block_id: BlockId) -> None:
-    """
-    Assert block_id is "latest" and throw RpcError otherwise
-    """
-    if block_id != "latest":
-        raise RpcError(code=-1, message="Calls with block_id != 'latest' are not supported currently.")
-
-
-def rpc_felt(value: int) -> str:
-    """
-    Convert integer to 0x0 prefixed felt
-    """
-    if value == 0:
-        return "0x00"
-    return "0x0" + hex(value).lstrip("0x")
-
-
-def rpc_root(root: str) -> str:
-    """
-    Convert 0 prefixed root to 0x prefixed root
-    """
-    root = root[1:]
-    return "0x0" + root
-
-
-def rpc_response(message_id: int, content: dict) -> dict:
-    """
-    Wrap response content in rpc format
-    """
-    return {
-        "jsonrpc": "2.0",
-        "id": message_id,
-        "result": content
-    }
-
-
-class RpcError(Exception):
-    """
-    Error message returned by rpc
-    """
-
-    def __init__(self, code, message):
-        super().__init__(message)
-        self.code = code
-        self.message = message
-
-
-def rpc_error(message_id: int, code: int, message: str) -> dict:
-    """
-    Wrap error in rpc format
-    """
-    return {
-        "jsonrpc": "2.0",
-        "id": message_id,
-        "error": {
-            "code": code,
-            "message": message
-        }
-    }
